backend/requirements.txt
# Core Flask Framework
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# AI & Machine Learning
google-generativeai==0.8.0

# HTTP & API
requests==2.31.0

# Environment Management
python-dotenv==1.0.0

# Optional: Production Server
# gunicorn==21.2.0

# Optional: Database (if needed)
# sqlalchemy==2.0.23
